64-bit filesize support
This patch is starting to get more interesting. It adds support for larger than
4GB filesizes (useful for e.g. large harddisk images or laserdisc videos).

This patch only changes the basic file-handling infrastructure, not yet the code
that actually uses this infrastructure.

Note that e.g. zip and gz files are still limited to 32-bit (because of the zlib
library and because of the zip file format).

More importantly, on 32 bit systems filesizes are still limited to 2GB. There's
no fundamental reason why we can't support larger files on 32 bit system. It's
just a lot more work than what this patch series does. Maybe we could still do
this in the future (though is it worth the effort?).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13085 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/cassette/CasImage.cc
@@ -152,7 +152,7 @@ bool CasImage::writeData(const byte* buf, const unsigned size, unsigned& pos)
 void CasImage::convert(const Filename& filename, FilePool& filePool, CliComm& cliComm)
 {
 	File file(filename);
-	unsigned size;
+	size_t size; // TODO
 	const byte* buf = file.mmap(size);
 
 	// search for a header in the .cas file
