+++ src/video/scalers/GLHQScaler.cc
@@ -67,7 +67,7 @@ GLHQScaler::GLHQScaler()
 		offsetTexture[i] = make_unique<Texture>();
 		offsetTexture[i]->setWrapMode(false);
 		offsetTexture[i]->bind();
-		unsigned size; // dummy
+		size_t size; // dummy
 		glTexImage2D(GL_TEXTURE_2D,       // target
 		             0,                   // level
 		             GL_RGBA8,            // internal format
